// src/models/ObjectMainServices.tsx
import React from 'react';

export class ObjectMainServices{

    iconOne: any;
    iconTwo: any;
    iconThree: any;
    iconFour: any;
    iconFive: any;
    iconSix: any;
    iconSeven: any;
    iconEight: any;
    iconNine: any;
    iconTen: any;
    iconEleven: any;
    iconTwelve: any;
    iconThirteen: any;
    iconFourteen: any;
    paypal: any;
    twists: any;
    idPay: any;

    constructor(){

    }
}